Added generic repository

+++ webApi/webApi/Domain/Entities/EmployeeProjectEntity.cs
@@ -4,12 +4,12 @@ using System.ComponentModel.DataAnnotations.Schema;
 namespace webApi.Domain.Entities
 {
     [Table("employees_projects")]
-    public class EmployeeProjectEntity
+    public class EmployeeProjectEntity 
     {
-        public int EmployeeId { get; set; }
+        public long EmployeeId { get; set; }
         public EmployeeEntity EmployeeEntity { get; set; }
 
-        public int ProjectId { get; set; }
+        public long ProjectId { get; set; }
         public ProjectEntity ProjectEntity { get; set; }
 
         public EmployeeProjectEntity()
